tidy: Add some explicit conversions to/from bool.
The clang-tidy "implicit boolean conversion" check pointed out a
number of places where booleans are implicitly converted to integers.
Make these conversion explicit.  The majority of these changes are
arguments to the QString::arg and QString::number functions.  Most
changes made by clang-tidy.

https://clang.llvm.org/extra/clang-tidy/checks/readability-implicit-bool-conversion.html

+++ mythtv/libs/libmyth/audio/audiooutputalsa.cpp
@@ -873,7 +873,7 @@ void AudioOutputALSA::SetVolumeChannel(int channel, int volume)
     auto chan = (snd_mixer_selem_channel_id_t) channel;
 
     if (snd_mixer_selem_has_playback_switch(m_mixer.elem))
-        snd_mixer_selem_set_playback_switch(m_mixer.elem, chan, (volume > 0));
+        snd_mixer_selem_set_playback_switch(m_mixer.elem, chan, static_cast<int>(volume > 0));
 
     if (snd_mixer_selem_set_playback_volume(m_mixer.elem, chan, mixervol) < 0)
         VBERROR(QString("failed to set channel %1 volume").arg(channel));
