+++ mythtv/libs/libmythbase/remotefile.cpp
@@ -171,7 +171,8 @@ MythSocket *RemoteFile::openSocket(bool control)
 
     if (control)
     {
-        strlist.append(QString("ANN Playback %1 %2").arg(hostname).arg(false));
+        strlist.append(QString("ANN Playback %1 %2")
+                       .arg(hostname).arg(static_cast<int>(false)));
         if (!lsock->SendReceiveStringList(strlist))
         {
             LOG(VB_GENERAL, LOG_ERR, loc +
@@ -184,8 +185,8 @@ MythSocket *RemoteFile::openSocket(bool control)
     else
     {
         strlist.push_back(QString("ANN FileTransfer %1 %2 %3 %4")
-                          .arg(hostname).arg(m_writeMode)
-                          .arg(m_useReadAhead).arg(m_timeoutMs));
+                          .arg(hostname).arg(static_cast<int>(m_writeMode))
+                          .arg(static_cast<int>(m_useReadAhead)).arg(m_timeoutMs));
         strlist << QString("%1").arg(dir);
         strlist << sgroup;
