+++ mythtv/libs/libmythtv/recorders/dvbchannel.cpp
@@ -974,7 +974,7 @@ int DVBChannel::GetChanID() const
 
     while (query.next())
     {
-        found += query.value(1).toInt() > 0;
+        found += static_cast<int>(query.value(1).toInt() > 0);
         if (id == -1 || found)
             id = query.value(0).toInt();
     }
